+++ app/src/main/java/com/example/schoolapp/datasource/online/model/ParentModel.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 import java.util.Date
 
 data class ParentModel(
-    @SerializedName("parentPhoneNumber") val parentPhoneNumber: Int,
+    @SerializedName("parentPhoneNumber") val parentPhoneNumber: Long,
     @SerializedName("parentJob") val parentJob: String,
     @SerializedName("parentNationality") val parentNationality: String,
     @SerializedName("parentAddress") val parentAddress: String
